% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_analysis.R
\name{checkGrainAge}
\alias{checkGrainAge}
\title{Check if grain ages are within a given standard deviation of a date}
\usage{
checkGrainAge(df, AgeField = "Age", SDField = "SD", checkAge, confidence = 2)
}
\arguments{
\item{df}{data frame}

\item{AgeField}{string, name of the Sample Age field}

\item{SDField}{string, name of the Sample Standard Deviation field}

\item{checkAge}{numeric, age of interest to check grain ages against}

\item{confidence}{numeric, number of standard deviations to check the grain age against}
}
\value{

}
\description{
Check if grain ages are within a given standard deviation of a date
}
