% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_analysis.R
\name{doseToAge}
\alias{doseToAge}
\title{Convert Paleodose and Sample Dose Rate to Age (before present)}
\usage{
doseToAge(DeGy, DoseRate)
}
\arguments{
\item{DeGy}{numeric or vector of numeric values of sample paleodose}

\item{DoseRate}{numeric or vector of numeric values of sample dose rate}
}
\value{
Age (numeric or vector of numeric values of sample age in years before sampled day)
}
\description{
Convert Paleodose and Sample Dose Rate to Age (before present)
}
\examples{
doseToAge(6.44, 2040)
}
