% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_analysis.R
\name{pct.younger}
\alias{pct.younger}
\title{Calculate the percent of a sample that is younger than a given date}
\usage{
pct.younger(
  df,
  AgeField = "ageGrid",
  DensField = "densities",
  IDField = "sample",
  yearBP = 1000
)
}
\arguments{
\item{df}{data frame of age densities}

\item{AgeField}{string, name of the Age field}

\item{DensField}{string, name of the Density field}

\item{IDField}{string, name of the Sample ID field}

\item{yearBP}{integer, value of the date to calculate percent younger than}
}
\value{
data frame containing percent younger than value for each sample
}
\description{
Calculate the percent of a sample that is younger than a given date
}
