% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_plots.R
\name{plot.OSLProfile}
\alias{plot.OSLProfile}
\title{plots a profile of OSL sample densities}
\usage{
\method{plot}{OSLProfile}(
  df,
  Xmin = 0,
  Xmax = 1e+05,
  densScale = 8,
  Ymin = 0,
  Ymax = 250,
  Ybreaks = seq(0, 300, 20),
  Xbreaks = c(seq(0, plotRange[2], 20000)),
  Xlabels = c(0, paste(seq(20000, plotRange[2], 20000)/1000, "k", sep = "")),
  vLine = 0,
  facetNo = 7
)
}
\arguments{
\item{df}{data frame, contains sample age-depth grids}

\item{Xmin}{integer, youngest age in years before present for the plot}

\item{Xmax}{integer, oldest age in years before present for the plot}

\item{densScale}{integer, value to scale the OSL densities on the plot}

\item{Ymin}{integer, shallowest depth for the plot}

\item{Ymax}{integer, deepest depth for the plot}

\item{Ybreaks}{vector, breaks for depths}

\item{Xbreaks}{vector, breaks for ages}

\item{Xlabels}{vector, labels for ages}

\item{vLine}{integer, value to draw a vertical red line showing the age of interest}

\item{facetNo}{integer, number of facets}
}
\value{
ggplot of OSL profile
}
\description{
plots a profile of OSL sample densities
}
