% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_plots.R
\name{plot_OSLProfile}
\alias{plot_OSLProfile}
\title{plots a profile of OSL sample densities}
\usage{
plot_OSLProfile(
  df,
  XRange = c(0, 1e+05),
  densScale = 8,
  YRange = c(0, 250),
  Ybreaks = seq(0, 300, 20),
  Xbreaks = c(seq(0, XRange[2], 20000)),
  Xlabels = c(0, paste(seq(20000, XRange[2], 20000)/1000, "k", sep = "")),
  vLine = 0
)
}
\arguments{
\item{df}{data frame, contains sample age-depth grids}

\item{XRange}{vector, minimum and maximum values of the x-axis}

\item{densScale}{integer, value to scale the OSL densities on the plot}

\item{YRange}{vector, minimum and maximum values of the y-axis}

\item{Ybreaks}{vector, breaks for depths}

\item{Xbreaks}{vector, breaks for ages}

\item{Xlabels}{vector, labels for ages}

\item{vLine}{integer, value to draw a vertical red line showing the age of interest}
}
\value{
ggplot of OSL profile
}
\description{
plots a profile of OSL sample densities
}
