% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_analysis.R
\name{sampleDensity}
\alias{sampleDensity}
\title{Convert sample ages to an age-density data frame}
\usage{
sampleDensity(
  df,
  AgeField = "Age",
  SDField = "SD",
  IDField = "Sample",
  grainSD = 2
)
}
\arguments{
\item{df}{data frame with sample results}

\item{AgeField}{string, name of the Sample Age field}

\item{SDField}{string, name of the Sample Age Standard Deviation field}

\item{IDField}{string, name of the Sample ID field}

\item{grainSD}{numeric greater than zero, value in standard deviations of to calculate the age-density grid}
}
\value{
data frame of age-densities
}
\description{
Convert sample ages to an age-density data frame
}
\examples{
myDF <- data.frame(Age = c(1000, 3000, 5000), SD = c(50, 40, 100), Sample = c('x1','x2','x3'))
ageDensities <- sampleDensity(myDF)
}
