% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osl_analysis.R
\name{single.osl}
\alias{single.osl}
\title{Fast sampling a normal distribution for plotting individual grain ages}
\usage{
single.osl(
  df,
  IDField,
  ID,
  AgeField = "Age",
  SDField = "SD",
  nSamples = 5000,
  checkAge = 0
)
}
\arguments{
\item{df}{data frame}

\item{IDField}{string, name of the Sample ID field}

\item{ID}{string, name of the sample ID to be analyzed}

\item{AgeField}{string, name of the Sample Age field}

\item{SDField}{string, name of the Sample Standard Deviation field}

\item{nSamples}{integer, number of samples to draw from a normal distribution}

\item{checkAge}{integer, age of sample to return T or F if the sample distribution is below or above a value}
}
\value{

}
\description{
Fast sampling a normal distribution for plotting individual grain ages
}
